@@ -213,7 +213,7 @@ func main() {
 	printResults(results, options, *serverStats)
 }
 
-func checkf(format string, args ...interface{}) {
+func checkf(format string, args ...any) {
 	for _, arg := range args {
 		if err, _ := arg.(error); err != nil {
 			fmt.Fprintf(os.Stderr, format, args...)
