@@ -211,7 +211,7 @@ func parseTemplate(name, tmpl string) *template.Template {
 	return template.Must(template.New(name).Funcs(funcMap).Parse(tmpl))
 }
 
-func mustExecute(t *template.Template, data interface{}) string {
+func mustExecute(t *template.Template, data any) string {
 	var b bytes.Buffer
 	if err := t.Execute(&b, data); err != nil {
 		panic(err)
