@@ -20,7 +20,7 @@ func TestSetup(t testing.TB, numServers int, srvFn func(i int) ServerIface) ([]s
 	t.Helper()
 	servers := make([]ServerIface, numServers)
 	addrs := make([]string, numServers)
-	for i := 0; i < numServers; i++ {
+	for i := range numServers {
 		srv := srvFn(i)
 		// listen on any available port
 		lis, err := net.Listen("tcp", "127.0.0.1:0")
