@@ -46,7 +46,7 @@ type asyncCallState struct {
 // This function should only be used by generated code.
 func (c RawConfiguration) AsyncCall(ctx context.Context, d QuorumCallData) *Async {
 	expectedReplies := len(c)
-	md := ordering.Metadata_builder{MessageID: c.getMsgID(), Method: d.Method}.Build()
+	md := ordering.NewGorumsMetadata(ctx, c.getMsgID(), d.Method)
 	replyChan := make(chan response, expectedReplies)
 
 	for _, n := range c {
