@@ -44,7 +44,7 @@ func TestChannelCreation(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	//the node should be closed manually because it isn't added to the configuration
+	// the node should be closed manually because it isn't added to the configuration
 	defer node.close()
 	mgr := dummyMgr()
 	// a proper connection should NOT be established here
@@ -52,8 +52,9 @@ func TestChannelCreation(t *testing.T) {
 
 	replyChan := make(chan response, 1)
 	go func() {
-		md := ordering.Metadata_builder{MessageID: 1, Method: handlerName}.Build()
-		req := request{ctx: context.Background(), msg: &Message{Metadata: md, Message: &mock.Request{}}}
+		ctx := context.Background()
+		md := ordering.NewGorumsMetadata(ctx, 1, handlerName)
+		req := request{ctx: ctx, msg: &Message{Metadata: md, Message: &mock.Request{}}}
 		node.channel.enqueue(req, replyChan, false)
 	}()
 	select {
@@ -118,7 +119,7 @@ func TestChannelReconnection(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	//the node should be closed manually because it isn't added to the configuration
+	// the node should be closed manually because it isn't added to the configuration
 	defer node.close()
 	mgr := dummyMgr()
 	// a proper connection should NOT be established here because server is not started
@@ -127,8 +128,9 @@ func TestChannelReconnection(t *testing.T) {
 	// send first message when server is down
 	replyChan1 := make(chan response, 1)
 	go func() {
-		md := ordering.Metadata_builder{MessageID: 1, Method: handlerName}.Build()
-		req := request{ctx: context.Background(), msg: &Message{Metadata: md, Message: &mock.Request{}}}
+		ctx := context.Background()
+		md := ordering.NewGorumsMetadata(ctx, 1, handlerName)
+		req := request{ctx: ctx, msg: &Message{Metadata: md, Message: &mock.Request{}}}
 		node.channel.enqueue(req, replyChan1, false)
 	}()
 
@@ -147,8 +149,9 @@ func TestChannelReconnection(t *testing.T) {
 	// send second message when server is up
 	replyChan2 := make(chan response, 1)
 	go func() {
-		md := ordering.Metadata_builder{MessageID: 2, Method: handlerName}.Build()
-		req := request{ctx: context.Background(), msg: &Message{Metadata: md, Message: &mock.Request{}}, opts: getCallOptions(E_Multicast, nil)}
+		ctx := context.Background()
+		md := ordering.NewGorumsMetadata(ctx, 2, handlerName)
+		req := request{ctx: ctx, msg: &Message{Metadata: md, Message: &mock.Request{}}, opts: getCallOptions(E_Multicast, nil)}
 		node.channel.enqueue(req, replyChan2, false)
 	}()
 
@@ -167,8 +170,9 @@ func TestChannelReconnection(t *testing.T) {
 	// send third message when server has been previously up but is now down
 	replyChan3 := make(chan response, 1)
 	go func() {
-		md := ordering.Metadata_builder{MessageID: 3, Method: handlerName}.Build()
-		req := request{ctx: context.Background(), msg: &Message{Metadata: md, Message: &mock.Request{}}}
+		ctx := context.Background()
+		md := ordering.NewGorumsMetadata(ctx, 3, handlerName)
+		req := request{ctx: ctx, msg: &Message{Metadata: md, Message: &mock.Request{}}}
 		node.channel.enqueue(req, replyChan3, false)
 	}()
 
