@@ -10,11 +10,11 @@ const (
 	// GenVersion is the runtime version required by generated .pb.go files.
 	// This is incremented when generated code relies on new functionality
 	// in the runtime.
-	GenVersion = 8
+	GenVersion = 9
 
 	// MinVersion is the minimum supported version for generated .pb.go files.
 	// This is incremented when the runtime drops support for old code.
-	MinVersion = 7
+	MinVersion = 9
 )
 
 // EnforceVersion is used by code generated by protoc-gen-gorums
